setnext official/smarty/requirements.txt
fastapi==0.109.2
langchain==0.1.7
langchain_core==0.1.23
langchain_openai==0.0.6
langchain_community==0.0.20
pinecone_client==3.0.3
redis==5.0.1
pydantic-settings==2.2.0
uvicorn